+++ src/main/scala/nn/trainers/gibbs/GibbsSample.scala
@@ -1,3 +1,3 @@
 package nn.trainers.gibbs
 
-case class GibbsSample(mean:Array[Double], sample:Array[Int])
+case class GibbsSample(mean:Array[Double], sample:Array[Double])
