+++ src/main/scala/nn/trainers/gibbs/GibbsSampler.scala
@@ -8,19 +8,19 @@ class GibbsSampler(rbm:RBM)(implicit rng:Random) {
   val numHidden = rbm.numHidden
   val numVisible = rbm.numVisible
 
-  def sampleGibbsHVH(h0Sample: Array[Int], nvMeans: Array[Double], nvSamples: Array[Int]): GibbsHVHSample = {
+  def sampleGibbsHVH(h0Sample: Array[Double], nvMeans: Array[Double], nvSamples: Array[Double]): GibbsHVHSample = {
     val vh = sampleVGivenH(h0Sample)
 
     GibbsHVHSample(vh, sampleHGivenV(vh.sample))
   }
 
-  def sampleHGivenV(v0Sample: Array[Int]): GibbsSample = {
+  def sampleHGivenV(v0Sample: Array[Double]): GibbsSample = {
     val mean = Range(0, numHidden).map { i => rbm.propagateUp(v0Sample, i) }.toArray
 
     GibbsSample(mean, sample(mean, rng))
   }
 
-  def sampleVGivenH(h0Sample: Array[Int]): GibbsSample = {
+  def sampleVGivenH(h0Sample: Array[Double]): GibbsSample = {
     val mean = Range(0, numVisible).map { i => rbm.propagateDown(h0Sample, i) }.toArray
 
     GibbsSample(mean, sample(mean, rng))
